Added HastingsFarm, moved JS into viewer.js, started working on container
