+++ BTWAD/hastings/demo.css
@@ -1,0 +1,3 @@
+.main-container {
+    display: flex;
+}
